moved helper functions to conftest to eliminate duplicate definitions (get_token and client)
