+++ backend/tests/conftest.py
@@ -30,3 +30,21 @@ def setup_database():
     SQLModel.metadata.create_all(test_engine)
     yield
     SQLModel.metadata.drop_all(test_engine)
+    
+
+from fastapi.testclient import TestClient
+from app.main import app
+
+client = TestClient(app)
+
+def get_token(email: str = "test@example.com") -> str:
+    client.post("/auth/register", json={
+        "email": email,
+        "password": "password123",
+        "full_name": "Test User"
+    })
+    response = client.post("/auth/login", json={
+        "email": email,
+        "password": "password123"
+    })
+    return response.json()["access_token"]
