+++ backend/tests/test_reports.py
@@ -1,26 +1,4 @@
-from fastapi.testclient import TestClient
-from app.main import app
-
-from sqlmodel import Session
-from app.models.currency import Currency
-from tests.conftest import test_engine
-from app.models.asset import Asset
-
-client = TestClient(app)
-
-
-def get_token(email: str = "test@example.com") -> str:
-    client.post("/auth/register", json={
-        "email": email,
-        "password": "password123",
-        "full_name": "Test User"
-    })
-    response = client.post("/auth/login", json={
-        "email": email,
-        "password": "password123"
-    })
-    return response.json()["access_token"]
-
+from tests.conftest import client, get_token
         
 def test_execute_buy_trade() -> None:
     token = get_token()
