+++ backend/tests/test_trade_execution.py
@@ -6,20 +6,7 @@ from app.models.currency import Currency
 from tests.conftest import test_engine
 from app.models.asset import Asset
 
-client = TestClient(app)
-
-
-def get_token(email: str = "test@example.com") -> str:
-    client.post("/auth/register", json={
-        "email": email,
-        "password": "password123",
-        "full_name": "Test User"
-    })
-    response = client.post("/auth/login", json={
-        "email": email,
-        "password": "password123"
-    })
-    return response.json()["access_token"]
+from tests.conftest import client, get_token
 
 def seed_asset(ticker: str = "AAPL") -> None:
     with Session(test_engine) as session:
@@ -61,7 +48,7 @@ def test_execute_sell_trade() -> None:
     token, account_id = create_trade_setup()
     
     # buy stock
-    response_buy = client.post(f"/portfolio/accounts/{account_id}", json={
+    client.post(f"/portfolio/accounts/{account_id}", json={
         "ticker": "AAPL",
         "direction": "buy",
         "quantity": 1
@@ -97,7 +84,7 @@ def test_sell_more_than_owned() -> None:
     num_sell = 2
     
     # buy stock
-    response_buy = client.post(f"/portfolio/accounts/{account_id}", json={
+    client.post(f"/portfolio/accounts/{account_id}", json={
         "ticker": "AAPL",
         "direction": "buy",
         "quantity": num_buy
@@ -110,6 +97,7 @@ def test_sell_more_than_owned() -> None:
         "quantity": num_sell
     }, headers={"Authorization": f"Bearer {token}"})
     
+    assert num_sell > num_buy
     assert response_sell.status_code == 400
 
 
@@ -126,11 +114,10 @@ def test_trade_wrong_account() -> None:
     }, headers={"Authorization": f"Bearer {token_1}"})
     account_id_1 = create_response_1.json()["id"]
 
-    create_response_2 = client.post("/accounts", json={
+    client.post("/accounts", json={
         "currency_code": "ZAR",
         "initial_balance": "1000.00"
     }, headers={"Authorization": f"Bearer {token_2}"})
-    account_id_2 = create_response_2.json()["id"]
     
     # trade with mismatched token
     response = client.post(f"/portfolio/accounts/{account_id_1}", json={
@@ -156,7 +143,7 @@ def test_get_holdings_after_buy() -> None:
     holdings_before = response_before_buy.json()["holdings"]
     net_quantity_before = holdings_before[0]["net_quantity"] if holdings_before else 0
     
-    response_buy = client.post(f"/portfolio/accounts/{account_id}", json={
+    client.post(f"/portfolio/accounts/{account_id}", json={
         "ticker": "AAPL",
         "direction": "buy",
         "quantity": 1
